Tweak scoring function
In fieldAccess, if a value is 0 it should not be replaced by the default value.
To avoid log(0) we translate stars * scalaPercentage by one.
The scoring function becomes:
log(stars * scalaPercentage + 1)
