--- modules/infra/src/main/scala/scaladex/infra/QueryDocumentFormat.scala
@@ -1,10 +1,0 @@
-package scaladex.infra
-
-trait QueryDocumentFormat {
-
-  def fieldAccess(name: String): String =
-    s"doc['$name'].value"
-
-  def fieldAccess(name: String, default: String): String =
-    s"((doc['$name'].size() != 0 && doc['$name'].value != 0) ? doc['$name'].value : $default)"
-}
